app/requirements.txt
Pillow==10.3.0
Flask==3.0.3
paho-mqtt==2.1.0
requests==2.31.0
